refactor: separate stack deployment
